+++ android/src/main/java/com/robinpowered/Frequency/FrequencyPackage.java
@@ -1,4 +1,4 @@
-package com.robinpowered.RNFrequency;
+package com.robinpowered.Frequency;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -9,10 +9,10 @@ import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
 import com.facebook.react.bridge.JavaScriptModule;
-public class RNFrequencyPackage implements ReactPackage {
+public class FrequencyPackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-      return Arrays.<NativeModule>asList(new RNFrequencyModule(reactContext));
+      return Arrays.<NativeModule>asList(new FrequencyModule(reactContext));
     }
 
     // Deprecated from RN 0.47
